Make AppMac::mImpl protected (formerly private)
It's currently necessary to subclass AppMac in order to subclass the obj-c AppImplMac bacuse mImpl is instantiated in AppMac's constructor. This change makes this possible by changing the access of mImpl to protected.

I'm curious why this was private to begin with. It might also be worth changing AppCocoaTouch and others.